# openapi.yml
openapi: 3.0.3
...
info:
  title: TooManyHours API
  description: |
    Game tracking API. Users keep a list across three categories and share it
    at a public profile URL.

    There is no document-level `security` default: every protected operation
    declares `BearerAuth` explicitly, so an operation without a `security`
    block is genuinely public.
  version: 0.2.0
servers:
  - url: http://localhost:3130
    description: >-
      Local development server. The frontend does not call this directly — Vite
      proxies /api/* here and strips the prefix.
...
paths:
...
  /authenticate:
    post:
      operationId: authenticate
      summary: Log in
      description: >-
        Returns 202 Accepted, not 200. Treating 200 as the success code would
        make a generated client reject every real login.
      tags: [Auth]
      security: []   # deliberately public
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AuthenticateRequest' }
      responses:
        '202':
          description: Authenticated; tokens returned and refresh cookie set
          headers:
            Set-Cookie:
              description: __Host-refresh_token; HttpOnly, Secure, SameSite=Strict
              schema: { type: string }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPairs' }
        '400':
          description: Malformed body, or invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '429':
          description: >-
            Too many failed attempts, either from this address (20 per 15
            minutes) or against this account (5 per minute). The body is
            identical whether or not the account exists.
          headers:
            Retry-After:
              description: Seconds to wait before attempting again
              schema: { type: integer }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Token generation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /refresh-token:
    get:
      operationId: refreshToken
      summary: Exchange the refresh cookie for a new token pair
      description: >-
        Authenticated by the `__Host-refresh_token` cookie rather than a bearer
        token, so the caller must send credentials with the request. The cookie
        carries only a token id; the server decides validity from its own
        records, which is what makes logout able to revoke it.


        Single use: each refresh consumes its token and returns a successor.
        Presenting an already-consumed token more than a few seconds later,
        while the session is still live, is treated as a leaked copy and
        revokes every token from that login.


        An access token is not accepted here. Access tokens carry no token id,
        and this endpoint requires one.
      tags: [Auth]
      security: []   # deliberately public
      responses:
        '200':
          description: New token pair issued and refresh cookie rotated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPairs' }
        '401':
          description: >-
            Cookie missing, unreadable, unknown, expired, already consumed,
            belonging to a logged-out session, or an access token presented in
            its place. Deliberately indistinguishable — the caller learns only
            that it failed. The response also expires the cookie.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Token generation failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /me:
    get:
      operationId: getMe
      summary: Get the authenticated user
      tags: [Users]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: The authenticated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: User lookup failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Missing, malformed, or expired bearer token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: User context missing from the request
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
    patch:
      operationId: patchMe
      summary: Update username and/or profile visibility
      description: >-
        At least one property must be present. The username is validated by the
        same rules as registration and stored lowercased.
      tags: [Users]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PatchMeRequest' }
      responses:
        '200':
          description: The updated user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: Empty body, invalid username, or invalid visibility
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Missing, malformed, or expired bearer token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          description: The user no longer exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: Username already taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Update failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /users/{id}:
    get:
      operationId: getUserById
      summary: Get a user by id
      tags: [Users]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: The user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '400':
          description: The id is not an integer
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401':
          description: Missing, malformed, or expired bearer token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Lookup failed, including when no such user exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  /genres:
    get:
      operationId: getGenres
      summary: List all genres
      tags: [Genres]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Genres ordered by name
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Tag' }
        '401':
          description: Missing, malformed, or expired bearer token
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500':
          description: Query failed
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        A 15-minute access token from /authenticate, /register or
        /refresh-token.
  schemas:
    Tag:
      type: object
      description: >-
        A genre, theme or game mode. IGDB attaches all three to a game and they
        behave identically, so they share one table and one type here.
      required: [id, name]
      properties:
        id: { type: integer, example: 7 }
        name: { type: string, example: Shooter }
...
    User:
      type: object
      required: [id, username, email, visibility, created_at, updated_at]
      properties:
        id: { type: integer }
        username:
          type: string
          description: Lowercase, unique, and the public profile URL key.
        email: { type: string, format: email }
        visibility:
          type: string
          enum: [public, private]
          description: >-
            Stored and editable, but not yet read by anything — public profiles
            are a later phase.
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TokenPairs:
      type: object
      required: [access_token, refresh_token]
      properties:
        access_token:
          type: string
          description: HS256 JWT, 15 minute expiry. Held in memory by the client.
        refresh_token:
          type: string
          description: >-
            HS256 JWT, 24 hour expiry. Also set as the __Host-refresh_token
            cookie; the frontend uses the cookie, not this copy.
...
    AuthenticateRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }
    PatchMeRequest:
      type: object
      description: >-
        At least one property must be present; an empty object is a 400.
      properties:
        username: { type: string, minLength: 3, maxLength: 16 }
        visibility: { type: string, enum: [public, private] }
...
    ErrorResponse:
      type: object
      description: >-
        The envelope produced by app.errorJSON, shared by every failure
        response.
      required: [error, message]
      properties:
        error: { type: boolean, example: true }
        message: { type: string }
        data:
          description: >-
            Declared by the Go struct but never populated by the current
            handlers.